Readme
The aim of This project is that it can detect COVID using deep learning. It makes use of CNN and utilizes Xray images as data sets. Xray images are fed into image classifying models and also segmentation techniques could be used on them to find affected parts. Normal as well as covid positive Xrays are fed into the model to train teh classifier. It elemenates the hard task to compare the Xrays , tackles the fine grained visual recognition problem as the Xrays are quite similar and need to be comapred on the basis of textures, sharpness etc heat map could highlight affected parts. images--->extract features--->classify model can also be trained for other pulmonary diseases similar in nature to covid like pneumonia , not only that the varient of pneumonia ie bacterial or viral can also be classified.